+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - Fix debug logging not logging if request was cached or not correctly
 - Add in SetupDialogComplete tracking event
 - Fix mods not behaving between MC version changes
+- Fix sort box on Add Mods dialog not showing